+++ client/src/components/LandingPage.js
@@ -1,5 +1,4 @@
-import React, { useEffect, useState } from "react";
-import { Link } from "react-router-dom";
+import React from "react";
 import SignInForm from "./authentication/SignInForm";
 
 const LandingPage = (props) => {
