+++ client/src/components/forms/AddPreferencesForm.js
@@ -1,13 +1,14 @@
 import React, { useState } from "react";
-import { Input, Text } from "@chakra-ui/react";
+import { Text } from "@chakra-ui/react";
 import CreatableSelect from "react-select/creatable";
-import options from "../services/userSelections";
 
-const EditPreferenceForm = ({ changePreference, preference, setEditForm }) => {
+import options from "../../services/userSelections";
+
+const AddPreferenceForm = (props) => {
     const [newPreference, setNewPreference] = useState({
         name: "",
+        icon: '<i class="fa-solid fa-user-plus"></i>',
     });
-
     const handleInputChange = (event) => {
         setNewPreference({
             ...newPreference,
@@ -17,42 +18,40 @@ const EditPreferenceForm = ({ changePreference, preference, setEditForm }) => {
 
     const handleSubmit = (event) => {
         event.preventDefault();
-        changePreference(newPreference);
+        props.onAddPreference(newPreference);
         clearForm();
     };
 
     const clearForm = () => {
         setNewPreference({
             name: "",
+            icon: '<i class="fa-solid fa-user-plus"></i>',
         });
-        setEditForm(false);
     };
-
     return (
         <div className="form">
             <form onSubmit={handleSubmit}>
                 <label htmlFor="name">
-                    <Text
-                        lg="20px"
-                        className="createable-select-title"
-                    >{`Edit Your ${preference.name} interest!`}</Text>
+                    <Text lg="20px" className="createable-select-title">
+                        Add a new Interest!
+                    </Text>
                     <CreatableSelect
                         isClearable
                         className="createable-select"
                         options={options}
                         onChange={handleInputChange}
-                        focusBorderColor="#13a200"
-                        placeholder={`Edit ${preference.name} here!`}
+                        focusBorderColor="#0606ff"
+                        placeholder="Type a new interest!"
                         defaultValue={newPreference.name}
                         size="sm"
                     />
                 </label>
                 <div className="button-add cell small-12" onClick={handleSubmit}>
-                    Edit
+                    Add
                 </div>
             </form>
         </div>
     );
 };
 
-export default EditPreferenceForm;
+export default AddPreferenceForm;
