+++ client/src/components/layout/ActivitySearchBar.js
