+++ client/src/components/layout/LocationSearchBar.js
