+++ client/src/components/listComponents/PopularDestinationList.js
@@ -1,6 +1,7 @@
 import React, { useEffect, useState } from "react";
-import PopularDestinationTile from "./PopularDestinationTile";
-import GetDestination from "../services/GetDestination";
+
+import PopularDestinationTile from "../tiles/PopularDestinationTile";
+import GetDestination from "../../services/GetDestination";
 
 const PopularDestinationList = (props) => {
     const [popularDestinations, setPopularDestinations] = useState([]);
