+++ client/src/components/tiles/ActivityTile.js
