+++ client/src/components/tiles/ResultTile.js
